% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{check_model_fn}
\alias{check_model_fn}
\title{Check that a decision-analytic model function is of the appropriate form.  Detect if it returns net benefit or CEA format
note check_outputs adds a class}
\usage{
check_model_fn(
  model_fn,
  par_fn,
  mfargs = NULL,
  outputs = NULL,
  verbose = FALSE
)
}
\value{
A modified copy of \code{model_fn} with a \code{class} attribute indicating whether it is
in net benefit \code{"nb"} or cost-effectiveness \code{"cea"} format, and an attribute \code{"nopt"} giving the
number of decision options.
}
\description{
Check that a decision-analytic model function is of the appropriate form.  Detect if it returns net benefit or CEA format
note check_outputs adds a class
}
\keyword{internal}
