% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_npreg.R
\name{check_regression}
\alias{check_regression}
\title{Check the fit of a regression model used to estimate EVPPI or EVSI}
\usage{
check_regression(x, pars = NULL, n = NULL, comparison = 1, outcome = "costs")
}
\arguments{
\item{x}{Output from \code{\link{evppi}} or \code{\link{evsi}}. The argument \code{check=TRUE}
must have been used when calling \code{evppi} or \code{evsi}, to allow the regression model
objects from \code{gam} or \code{earth} to be preserved.  (This is not done by
default, since these objects can be large.)}

\item{pars}{Parameter (or parameter group) whose EVPPI calculation is to be checked.
This should be in the \code{pars} component of the object returned by \code{\link{evppi}}.
Only relevant if \code{x} is the result of an \code{\link{evppi}} calculation.}

\item{n}{Sample size whose EVSI calculation is to be checked.
This should be in the \code{n} component of the object returned by \code{\link{evsi}}.
Only relevant if \code{x} is the result of an \code{\link{evsi}} calculation.}

\item{comparison}{Only relevant if there are more than two treatments in the decision model.
Different regression models are then used for the comparisons of different treatments
with the baseline treatment.
\code{comparison} is an integer identifying which of these models is checked.}

\item{outcome}{\code{"costs"} or \code{"effects"}.  Only relevant if \code{outputs} was
in cost-effectiveness format when
calling \code{evppi} or \code{evsi}, hence different regressions are used for costs and
effects.  By default, \code{outcome="costs"} is used, so that the regression
for costs is checked.}
}
\value{
There is no return value.  Plots are produced, and outputs are
printed.
}
\description{
Uses the standard checking methods built into the \code{mgcv} and \code{earth} packages,
which produce plots of the residuals from the regression.
}
\details{
For VoI estimation, the key thing we are looking for is that the residuals
have mean zero, hence that the mean of the model output is represented well by the
regression function of the model input parameters.  It should not matter if the
variance of the residuals is
non-constant, or non-normally distributed.

Models produced with \code{method="gam"} are checked using \code{\link{gam.check}}

Models produced \code{method="earth"} are checked using \code{\link{plot.earth}}
}
\examples{
# TODO and refer in vignette

}
