% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemo_cea}
\alias{chemo_cea}
\alias{chemo_nb}
\alias{chemo_pars}
\title{Chemotherapy health economic model}
\format{
List

An object of class \code{matrix} (inherits from \code{array}) with 10000 rows and 2 columns.

An object of class \code{data.frame} with 10000 rows and 23 columns.
}
\source{
TODO
}
\usage{
chemo_cea

chemo_nb

chemo_pars
}
\description{
Chemotherapy health economic model
}
\references{
Convoi1
}
\keyword{datasets}
