% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_chemo.R
\name{chemo_costeff_fn}
\alias{chemo_costeff_fn}
\title{Chemotherapy health economic model}
\usage{
chemo_costeff_fn(SE1, SE2, lambda.home.home, lambda.home.hosp,
  lambda.home.rec, lambda.hosp.hosp, lambda.hosp.rec, lambda.hosp.dead,
  c.home, c.hosp, c.dead, e.chemo, e.home, e.hosp)
}
\arguments{
\item{SE1}{TODO}

\item{SE2}{TODO}

\item{lambda.home.home}{Transition probability from}

\item{lambda.home.hosp}{Transition probability from}

\item{lambda.home.rec}{Transition probability from}

\item{lambda.hosp.hosp}{Transition probability from}

\item{lambda.hosp.rec}{Transition probability from}

\item{lambda.hosp.dead}{Transition probability from}

\item{c.home}{Cost TODO}

\item{c.hosp}{Cost TODO}

\item{c.dead}{Cost TODO}

\item{e.chemo}{Effect TODO}

\item{e.home}{Effect TODO}

\item{e.hosp}{Effect TODO}
}
\description{
Calculate expected costs and effects for two treatments as a function of the input parameters.   Each input parameter is a scalar.
}
\details{
TODO refer to example of looping to generate PSA output matrix (currently in data-raw)
}
