% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_chemo_wrapper.R
\name{chemo_model}
\alias{chemo_model}
\alias{chemo_pars_fn}
\alias{chemo_model_nb}
\alias{chemo_model_cea}
\title{Chemotherapy cost-effectiveness model}
\usage{
chemo_pars_fn(n)

chemo_model_nb(
  p_side_effects_t1,
  p_side_effects_t2,
  p_home_home,
  p_home_hospital,
  p_home_recover,
  p_hospital_hospital,
  p_hospital_recover,
  p_hospital_dead,
  c_home_care,
  c_hospital,
  c_death,
  u_recovery,
  u_home_care,
  u_hospital
)

chemo_model_cea(
  p_side_effects_t1,
  p_side_effects_t2,
  p_home_home,
  p_home_hospital,
  p_home_recover,
  p_hospital_hospital,
  p_hospital_recover,
  p_hospital_dead,
  c_home_care,
  c_hospital,
  c_death,
  u_recovery,
  u_home_care,
  u_hospital
)
}
\arguments{
\item{p_side_effects_t1}{Probability of side effects under treatment 1}

\item{p_side_effects_t2}{Probability of side effects under treatment 2}

\item{p_home_home}{Annual transition probability from home to home}

\item{p_home_hospital}{Transition probability from home to hospital}

\item{p_home_recover}{Transition probability from home to recovery}

\item{p_hospital_hospital}{Transition probability from hospital to hospital}

\item{p_hospital_recover}{Transition probability from hospital to recovery}

\item{p_hospital_dead}{Transition probability from hospital to death}

\item{c_home_care}{Cost of a yearly period under treatment at home}

\item{c_hospital}{Cost of hospital treatment}

\item{c_death}{Cost of death}

\item{u_recovery}{Utility of a period in the recovery state}

\item{u_home_care}{Utility of home care state}

\item{u_hospital}{Utility of hospital state}
}
\value{
\code{chemo_model_nb} returns a vector with elements giving the net monetary benefit for standard of care
and novel treatment, respectively, at a willingness-to-pay of 20,000 pounds per QALY.

\code{chemo_model_cea} returns a matrix with:
\itemize{
\item two rows, the first for expected costs and the second for expected effects (QALYs) over the fifty year time horizon, and
\item two columns, the first for the "standard of care" decision option, and the second for the novel
treatment.
}

\code{chemo_pars_fn} generates a sample from the uncertainty distribution of the parameters in the chemotherapy model . This returns a data frame with parameters matching the arguments of
\code{\link{chemo_model_nb}}.  Additional derived parameters are:
\itemize{
\item \code{rr_side_effects}: relative risk of side effects for treatment 2 compared to 1
\item \code{p_hospitalised_total}: probability of hospitalisation over the 50 year time horizon
\item \code{p_died}: probability of death over the time horizon, given hospitalisation
\item \code{lambda_home}: conditional probability that a patient recovers given they are not hospitalised
\item \code{lambda_hosp}: conditional probability that a patient in hospital recovers given they do not die
}
}
\description{
An artificial health economic decision model with a typical Markov model structure, used for illustrating Value of Information methods.
}
