% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_chemo_old.R
\name{chemo_prior_pars}
\alias{chemo_prior_pars}
\title{Chemotherapy health economic model parameters}
\usage{
chemo_prior_pars(n)
}
\arguments{
\item{n}{Number of samples to draw}
}
\description{
Generate a sample from the uncertainty distribution of the
parameters.
}
\details{
This is the state of knowledge under current information, prior to
any further research.  The distributions are all of standard
analytic forms.  The distributions of pi1, gamma.hosp and
gamma.dead are obtained from simple beta/binomial conjugate
Bayesian inference based on current data.
}
