% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_so.R
\name{form_nbarray}
\alias{form_nbarray}
\title{Convert either "net benefit" or "cost-effectiveness" forms for
'outputs' to a 3D array  (number of simulations, x number of
willingness-to-pay values, x number of decision options) with one
WTP value when \code{outputs} is in "net benefit" form.}
\usage{
form_nbarray(outputs)
}
\description{
Convert either "net benefit" or "cost-effectiveness" forms for
'outputs' to a 3D array  (number of simulations, x number of
willingness-to-pay values, x number of decision options) with one
WTP value when \code{outputs} is in "net benefit" form.
}
\keyword{internal}
