% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.evppi}
\alias{plot.evppi}
\title{Plot EVPPI estimates as a simple dot plot}
\usage{
\method{plot}{evppi}(x, order = FALSE, wtp = NULL, ...)
}
\arguments{
\item{x}{Object returned from \code{\link{evppi}}.}

\item{order}{Order the plot with highest EVPPI values at the top.}

\item{wtp}{Not implemented yet. Suggestions welcome for how. Multiple curves on same plot, for top few
or selected?}

\item{...}{Other arguments (currently unused).}
}
\value{
A \code{ggplot2} object.
}
\description{
This is a simple \code{ggplot2} plot and returns a \code{ggplot2} object, which
can be manipulated.  Or examine the source for this function to see how
to construct a similar plot that can be customised.
}
