% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemo_cea}
\alias{chemo_cea}
\alias{chemo_nb}
\alias{chemo_pars}
\title{Chemotherapy health economic model}
\format{List}
\source{
TODO
}
\usage{
chemo_cea

chemo_nb

chemo_pars
}
\description{
Chemotherapy health economic model
}
\references{
Convoi1
}
\keyword{datasets}
