% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{evsi}
\alias{evsi}
\title{Calculate the expected value of sample information under a decision-analytic model}
\usage{
evsi(outputs, inputs, rfn, method, ...)
}
\arguments{
\item{outputs}{This could take one of two forms

"net benefit" form: a matrix or data frame of samples from the uncertainty distribution of the expected net benefit.  The number of rows should equal the number of samples, and the number of columns should equal the number of decision options.

"cost-effectiveness analysis" form: a list with the following named components:

\code{"c"}: a matrix or data frame of samples from the distribution of costs.  There should be one column for each decision option.

\code{"e"}: a matrix or data frame of samples from the distribution of effects, likewise.

\code{"k"}: a vector of willingness-to-pay values.

Objects of class \code{"bcea"}, as created by the \pkg{BCEA} package, may be supplied as the \code{outputs} argument, as they are lists with the required components. 

If \code{outputs} is a matrix or data frame it is assumed to be of "net benefit" form.  Otherwise if it is a list, it is assumed to be of "cost effectiveness analysis" form.

The exact names and formats of all these arguments is up for discussion!}

\item{inputs}{Matrix or data frame of samples from the uncertainty distribution of the input parameters of the decision model.   The number of columns should equal the number of parameters, and the columns should be named.    This should have the same number of rows as there are samples in \code{outputs}, and each row of the samples in \code{outputs} should give the model output evaluated at the corresponding parameters.}

\item{rfn}{Function to sample data from a proposed future study}

\item{method}{Character string indicating the calculation method.

Should support Strong, Heath, Jalal, Menzies at least (but named after method not person, as in convoi 1 paper)}

\item{...}{Other arguments required by specific methods}
}
\description{
Calculate the expected value of sample information under a decision-analytic model
}
