% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{evsi}
\alias{evsi}
\title{Calculate the expected value of sample information from a decision-analytic model}
\usage{
evsi(
  outputs,
  inputs,
  rfn,
  n = 100,
  method = NULL,
  likelihood = NULL,
  poi = NULL,
  npreg_method = "gam",
  nsim = NULL,
  ...
)
}
\arguments{
\item{outputs}{This could take one of two forms

"net benefit" form: a matrix or data frame of samples from the uncertainty distribution of the expected net benefit.  The number of rows should equal the number of samples, and the number of columns should equal the number of decision options.

"cost-effectiveness analysis" form: a list with the following named components:

\code{"c"}: a matrix or data frame of samples from the distribution of costs.  There should be one column for each decision option.

\code{"e"}: a matrix or data frame of samples from the distribution of effects, likewise.

\code{"k"}: a vector of willingness-to-pay values.

Objects of class \code{"bcea"}, as created by the \pkg{BCEA} package, may be supplied as the \code{outputs} argument, as they are lists with the required components. 

If \code{outputs} is a matrix or data frame it is assumed to be of "net benefit" form.  Otherwise if it is a list, it is assumed to be of "cost effectiveness analysis" form.

The exact names and formats of all these arguments is up for discussion!}

\item{inputs}{Matrix or data frame of samples from the uncertainty distribution of the input parameters of the decision model.   The number of columns should equal the number of parameters, and the columns should be named.    This should have the same number of rows as there are samples in \code{outputs}, and each row of the samples in \code{outputs} should give the model output evaluated at the corresponding parameters.}

\item{rfn}{Function to sample predicted data from a proposed future study.

This should have first argument defined by a data frame of parameter simulations, with one row per simulation and one column per parameter.  The names of the parameters must all correspond to parameters in \code{inputs}.

The function should return a data frame with the number of rows equal to the number of parameter simulations.  Thus if \code{inputs} is considered as a sample from the posterior, then \code{rfn} returns a corresponding sample from the posterior predictive distribution, which includes two sources of uncertainty: (a) uncertainty about the parameters and (b) sampling variation in observed data given fixed parameter values.}

\item{n}{Sample size of future study - optional argument to rfn - facilitates calculating EVSI for multiple sample sizes.  TODO how will this work for randomised trials with multiple arms?}

\item{method}{Character string indicating the calculation method.

All the nonparametric regression methods supported for \code{\link{evppi}}, that is \code{"gam","gp","earth","inla"}, can also be used for EVSI calculation by regressing on a summary statistics of the predicted data (Strong et al 201?).   Defaults to \code{"gam"}.

\code{"is"} for importance sampling (Menzies 200?)

TODO Heath (moment matching \code{"mm"}) and Jalal methods}

\item{likelihood}{Likelihood function, required (and only required) for the importance sampling method.  This should take arguments:

the first: a data frame with columns defined by the number of outcomes in the data, and with names matching the names of the data frame returned by \code{rfn}. 

the second: a data frame of parameter values, whose names should all correspond to variables in \code{inputs}.

The function should return a vector whose length matches the number of rows of the data frame given as the second argument. 

POINT TO AN EXAMPLE WHICH WILL MAKE ALL THIS CLEARER

Note the definition of the likelihood should agree with the definition of \code{rfn} to define a consistent sampling distribution for the data.   CLARIFY.  [ eventually we'll want some built-in common examples where people don't have to specify either rfn or likelihood ]}

\item{poi}{Parameters of interest, that is, those which are informed by the data in the future study.  Required (and only required) for the methods which involve an intermediate EVPPI calculation, that is the \code{"is"} and TODO OTHER methods.

This should bee a character vector naming particular columns of \code{inputs}.  It should consist of the variables used in the definition of \code{rfn} (and \code{likelihood} if used) and only these variables.}

\item{npreg_method}{Method to use to calculate the EVPPI, for those methods that require it.    STATE SUPPORTED VALUES}

\item{nsim}{Number of simulations from the model to use for calculating EVPPI.  The first \code{nsim} rows of the objects in \code{inputs} and \code{outputs} are used.}

\item{...}{Other arguments required by specific methods}
}
\description{
Calculate the expected value of sample information from a decision-analytic model
}
