% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{evsi}
\alias{evsi}
\title{Calculate the expected value of sample information under a decision-analytic model}
\usage{
evsi(nb, pars, rfn, method, ...)
}
\arguments{
\item{nb}{Matrix or data frame of samples from the uncertainty distribution of the expected net benefit.  The number of rows should equal the number of samples, and the number of columns should equal the number of decision options. 

The exact names of all these arguments is up for discussion! 

(if we want to transform the NB to the INB to make certain calculation methods more efficient, we can do that internally. this would need a "baseline" treatment.  would the user need to supply this?  should default to first column at least)

Should also be easy to compute for different willingness to pay values.  So should think about how to separate out costs and effects}

\item{pars}{Matrix or data frame of samples from the uncertainty distribution of the parameters of the decision-analytic model.   The number of columns should equal the number of parameters, and the columns should be named.    This should have the same number of rows as \code{nb}, and each row of \code{nb} should give the net benefit function evaluated at the corresponding parameters.}

\item{rfn}{Function to sample data from a proposed future study}

\item{method}{Character string indicating the calculation method.

Should support Strong, Heath, Jalal, Menzies at least (but named after method not person, as in convoi 1 paper)}

\item{...}{Other arguments required by specific methods}
}
\description{
Calculate the expected value of sample information under a decision-analytic model
}
